Move program back to original location
Prior to this change, the program was available only at a peculiar URL,
and was shown as a Pretalx embed on a page that featured CSS from the
home page, but with incorrect fonts in places, layout issues, and
missing navigation components.

This change retains the Pretalx embed, but moves it back to the expected
location for the schedule. It also replaces the custom HTML with the
base page Astro component, restoring the correct fonts and navigation,
and the correct styles for a page that isn't the home page.