Begin factoring out schedule into components
Prior to this change, the schedule was one big ol' component, because it
was ported across from a previous SSG which didn't offer components.
This means that we can't take full advantage of the modularity and
encapsulation components offer, and customising the appearance of the
schedule is currently complicated because of the need to wade through a
bunch of unrelated Astro and CSS code.

This change separates out the Session component, which is the individual
cells that appear in the schedule. There is still more refactoring to be
done but this is a start.